requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.9.1
aiosignal==1.3.1
amqp==5.2.0
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==25.1.0
backports.zoneinfo==0.2.1
bcrypt==4.2.0
beautifulsoup4==4.13.4
billiard==4.2.1
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.1
cryptography==43.0.1
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.8.0rc2
Django==4.2.15
django-allauth==64.1.0
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-db-multitenant==0.3.2
django-filter==24.3
django-js-asset==2.2.0
django-leaflet==0.31.0
django-microsoft-auth==3.0.1
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
django-tenant-schemas==1.12.0
django-tenants==3.7.0
django-webpush==0.3.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
ecdsa==0.19.1
et-xmlfile==1.1.0
firebase-admin==6.6.0
fonttools==4.54.1
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
html5lib==1.1
http_ece==1.2.1
httplib2==0.22.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.5
kiwisolver==1.4.7
kombu==5.4.2
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.7.5
msal==1.31.1
msgpack==1.1.0
mssql-django==1.5
multidict==6.1.0
ngrok==1.4.0
num2words==0.5.13
numpy==1.24.4
oauthlib==3.2.2
openpyxl==3.1.5
ordered-set==4.1.0
oscrypto==1.3.0
packaging==24.2
pandas==2.0.3
pdfkit==1.0.0
pillow==10.4.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.3
psycopg==3.2.4
psycopg2-binary==2.9.10
py-vapid==1.9.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.11.0
pyee==11.1.1
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
PyMuPDF==1.24.11
PyMySQL==1.1.1
pyodbc==5.1.0
pyparsing==3.1.4
pypdf==5.0.0
PyPDF2==3.0.1
pyphen==0.16.0
pypng==0.20220715.0
pyppeteer==2.0.0
python-barcode==0.15.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2024.1
pywebpush==2.0.3
PyYAML==6.0.2
qrcode==7.4.2
razorpay==1.4.1
redis==5.2.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sendgrid==6.12.2
sib-api-v3-sdk==7.6.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.7
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
tqdm==4.67.1
twilio==9.4.6
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==1.26.20
vine==5.1.0
waitress==3.0.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.8
wfastcgi==3.0.0
whitenoise==6.7.0
xhtml2pdf==0.2.16
yarl==1.15.2
zipp==3.20.2
zopfli==0.2.3
